refactor: rename UploadCache to WriteCache
