+++ src/mito2/src/cache.rs
@@ -18,7 +18,7 @@ mod cache_size;
 #[cfg(test)]
 pub(crate) mod test_util;
 #[allow(unused)]
-pub(crate) mod upload_cache;
+pub(crate) mod write_cache;
 
 use std::mem;
 use std::sync::Arc;
@@ -31,7 +31,7 @@ use parquet::file::metadata::ParquetMetaData;
 use store_api::storage::RegionId;
 
 use crate::cache::cache_size::parquet_meta_size;
-use crate::cache::upload_cache::UploadCacheRef;
+use crate::cache::write_cache::WriteCacheRef;
 use crate::metrics::{CACHE_BYTES, CACHE_HIT, CACHE_MISS};
 use crate::sst::file::FileId;
 
@@ -52,10 +52,10 @@ pub struct CacheManager {
     vector_cache: Option<VectorCache>,
     /// Cache for SST pages.
     page_cache: Option<PageCache>,
-    /// Upload cache.
-    // TODO(yingwen): Remove this once the upload cache is ready.
+    /// A Cache for writing files to object stores.
+    // TODO(yingwen): Remove this once the cache is ready.
     #[allow(unused)]
-    upload_cache: Option<UploadCacheRef>,
+    write_cache: Option<WriteCacheRef>,
 }
 
 pub type CacheManagerRef = Arc<CacheManager>;
@@ -115,7 +115,7 @@ impl CacheManager {
             sst_meta_cache,
             vector_cache,
             page_cache,
-            upload_cache: None,
+            write_cache: None,
         }
     }
