+++ src/mito2/src/cache/write_cache.rs
@@ -34,18 +34,18 @@ use crate::wal::EntryId;
 /// A cache for uploading files to remote object stores.
 ///
 /// It keeps files in local disk and sends files to object store in background.
-pub(crate) struct UploadCache {
+pub(crate) struct WriteCache {
     /// Local object storage to store files to upload.
     local_store: ObjectStore,
     /// Object store manager.
     object_store_manager: ObjectStoreManagerRef,
 }
 
-pub(crate) type UploadCacheRef = Arc<UploadCache>;
+pub(crate) type WriteCacheRef = Arc<WriteCache>;
 
-impl UploadCache {
+impl WriteCache {
     // TODO(yingwen): Maybe pass cache path instead of local store.
-    /// Create the upload cache with a `local_store` to cache files and a
+    /// Create the cache with a `local_store` to cache files and a
     /// `object_store_manager` for all object stores.
     pub(crate) fn new(
         local_store: ObjectStore,
